画面遷移の制御とrefactor
