--- src/Assets/Scripts/Core/Navigation/Default.cs
@@ -1,4 +1,0 @@
-// namespace System.Runtime.CompilerServices
-// {
-//     internal static class IsExternalInit {}
-// }
