--- src/Assets/Scripts/Core/Navigation/Default.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 356ec13e272641e3b4bd09fe32bf2560
-timeCreated: 1730042611
