+++ src/Assets/Scripts/Core/Utilities/Exception.cs
@@ -1,0 +1,14 @@
+namespace Core.Utilities
+{
+    public static class Exception
+    {
+        public static T CheckNull<T>(this T obj)
+        {
+            if (obj == null)
+            {
+                throw new System.ArgumentNullException(nameof(obj));
+            }
+            return obj;
+        }
+    }
+}
