+++ src/Assets/Scripts/Core/Utilities/Exception.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: f428f7bea4ec464ca13e82ba821e4237
+timeCreated: 1730703328
