+++ src/Assets/Scripts/Feature/Interface/IInputController.cs
@@ -1,7 +1,13 @@
 namespace Feature.Interface
 {
+    /// <summary>
+    /// 入力制御のためのインターフェース
+    /// </summary>
     public interface IInputController
     {
+        /// <summary>
+        /// 入力制御の初期化を行います
+        /// </summary>
         public void Start();
     }
 }
