+++ src/Assets/Scripts/Main/Controller/GameNavigation/AScreenProtocolLogic.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 53b04f68b78a46579db3632bf4eaa4b5
+timeCreated: 1730703007
